fixed and improved positioning of sort options (#72)
Fixed a bug in IE11 where the site was shifting offscreen when clicking a facet. The cause was the Foundation 'Action Sheet' CSS used for the sort menu component.

There was a secondary issue where the sort options were positioned incorrectly on all browsers and half of the options element was offscreen.

This PR fixes both issues.